yaml_rust2::char_traits
Function is_valid_literal_block_scalar

pub(crate) fn is_valid_literal_block_scalar(string: &str) -> bool

Check if the string can be expressed a valid literal block scalar. The YAML spec supports all of the following in block literals except #xFEFF:

    #x9 | #xA | [#x20-#x7E]                /* 8 bit */
  | #x85 | [#xA0-#xD7FF] | [#xE000-#xFFFD] /* 16 bit */
  | [#x10000-#x10FFFF]                     /* 32 bit */