yaml_rust2

Module parser

Home to the YAML Parser.

The parser takes input from the crate::scanner::Scanner, performs final checks for YAML compliance, and emits a stream of tokens that can be used by the crate::YamlLoader to construct the crate::Yaml object.

Structs

Parser

A YAML parser.

Tag

A YAML tag.

Enums

Event

An event generated by the YAML parser.

Traits

EventReceiver

Trait to be implemented in order to use the low-level parsing API.

MarkedEventReceiver

Trait to be implemented for using the low-level parsing API.

Type Aliases

ParseResult

A convenience alias for a Result of a parser event.