Function yaml_rust2::yaml::detect_utf16_endianness

fn detect_utf16_endianness(b: &[u8]) -> &'static Encoding

The encoding crate knows how to tell apart UTF-8 from UTF-16LE and utf-16BE, when the bytestream starts with BOM codepoint. However, it doesn’t even attempt to guess the UTF-16 endianness of the input bytestream since in the general case the bytestream could start with a codepoint that uses both bytes.

The YAML-1.2 spec mandates that the first character of a YAML document is an ASCII character. This allows the encoding to be deduced by the pattern of null (#x00) characters. See spec at https://yaml.org/spec/1.2/spec.html#id2771184