Module yaml_rust2::char_traits

Holds functions to determine if a character belongs to a specific character set.

Functions

as_hex ğŸ”’

Convert the hexadecimal digit to an integer.

is_alpha ğŸ”’

Check whether the character is a digit, letter, _ or -.

is_anchor_char ğŸ”’

Check whether the character is a valid YAML anchor name character.

is_blank ğŸ”’

Check whether the character is a whitespace ( or \t).

is_blank_or_breakz ğŸ”’

Check whether the character is nil, a linebreak or a whitespace.

is_bom ğŸ”’

Check whether the character is the BOM character.

is_break ğŸ”’

Check whether the character is a line break (\r or \n).

is_breakz ğŸ”’

Check whether the character is nil or a line break (\0, \r, \n).

is_digit ğŸ”’

Check whether the character is an ascii digit.

is_flow ğŸ”’

Check whether the character is a YAML flow character (one of ,[]{}).

is_hex ğŸ”’

Check whether the character is a hexadecimal character (case insensitive).

is_tag_char ğŸ”’

Check whether the character is a valid tag character.

is_uri_char ğŸ”’

Check whether the character is a valid URI character.

is_valid_literal_block_scalar ğŸ”’

Check if the string can be expressed a valid literal block scalar. The YAML spec supports all of the following in block literals except #xFEFF:

is_word_char ğŸ”’

Check whether the character is a valid word character.

is_yaml_non_break ğŸ”’

Check whether the character is a YAML non-breaking character.

is_yaml_non_space ğŸ”’

Check whether the character is NOT a YAML whitespace ( / \t).

is_z ğŸ”’

Check whether the character is nil (\0).