Module yaml_rust2::scanner

Home to the YAML Scanner.

The scanner is the lowest-level parsing utility. It is the lexer / tokenizer, reading input a character at a time and emitting tokens that can later be interpreted by the crate::parser to check for more context and validity.

Due to the grammar of YAML, the scanner has to have some context and is not error-free.

Structs

Indent 🔒

An indentation level on the stack of indentations.

Marker

A location in a yaml document.

ScanError

An error that occurred while scanning.

Scanner

The YAML scanner.

SimpleKey 🔒

A scalar that was parsed and may correspond to a simple key.

Token

A scanner token.

Enums

Chomping

Chomping, how final line breaks and trailing empty lines are interpreted.

SkipTabs 🔒

Behavior to adopt regarding treating tabs as whitespace.

TEncoding

The encoding of the input. Currently, only UTF-8 is supported.

TScalarStyle

The style as which the scalar was written in the YAML document.

TokenType

The contents of a scanner token.

Constants

BUFFER_LEN 🔒

The size of the Scanner buffer.

Type Aliases

ScanResult

A convenience alias for scanner functions that may fail without returning a value.