yaml_rust2
Module scanner

Home to the YAML Scanner.

The scanner is the lowest-level parsing utility. It is the lexer / tokenizer, reading input a character at a time and emitting tokens that can later be interpreted by the crate::parser to check for more context and validity.

Due to the grammar of YAML, the scanner has to have some context and is not error-free.

Structs

Marker

A location in a yaml document.

ScanError

An error that occurred while scanning.

Scanner

The YAML scanner.

Token

A scanner token.

Enums

Chomping

Chomping, how final line breaks and trailing empty lines are interpreted.

TEncoding

The encoding of the input. Currently, only UTF-8 is supported.

TScalarStyle

The style as which the scalar was written in the YAML document.

TokenType

The contents of a scanner token.

Type Aliases

ScanResult

A convenience alias for scanner functions that may fail without returning a value.