yaml_rust2
Module yaml

YAML objects manipulation utilities.

Macros

define_as 🔒

define_as_mut_ref 🔒

define_as_ref 🔒

define_into 🔒

Structs

YamlDecoder

YamlDecoder is a YamlLoader builder that allows you to supply your own encoding error trap. For example, to read a YAML file while ignoring Unicode decoding errors you can set the encoding_trap to encoding::DecoderTrap::Ignore.

YamlIter

An iterator over a Yaml node.

YamlLoader

Main structure for quickly parsing YAML.

Enums

LoadError

An error that happened when loading a YAML document.

YAMLDecodingTrap

The behavior YamlDecoder must have when an decoding error occurs.

Yaml

A YAML node is stored as this Yaml enumeration, which provides an easy way to access your YAML document.

Statics

BAD_VALUE 🔒

Functions

decode_loop 🔒

Perform a loop of Decoder::decode_to_string, reallocating output if needed.

detect_utf16_endianness 🔒

The encoding crate knows how to tell apart UTF-8 from UTF-16LE and utf-16BE, when the bytestream starts with BOM codepoint. However, it doesn’t even attempt to guess the UTF-16 endianness of the input bytestream since in the general case the bytestream could start with a codepoint that uses both bytes.

parse_f64 🔒

Type Aliases

Array

The type contained in the Yaml::Array variant. This corresponds to YAML sequences.

Hash

The type contained in the Yaml::Hash variant. This corresponds to YAML mappings.

YAMLDecodingTrapFn

The signature of the function to call when using YAMLDecodingTrap::Call.