yaml_rust2

Module yaml

YAML objects manipulation utilities.

Structs

YamlDecoder

YamlDecoder is a YamlLoader builder that allows you to supply your own encoding error trap. For example, to read a YAML file while ignoring Unicode decoding errors you can set the encoding_trap to encoding::DecoderTrap::Ignore.

YamlIter

An iterator over a Yaml node.

YamlLoader

Main structure for quickly parsing YAML.

Enums

LoadError

An error that happened when loading a YAML document.

YAMLDecodingTrap

The behavior YamlDecoder must have when an decoding error occurs.

Yaml

A YAML node is stored as this Yaml enumeration, which provides an easy way to access your YAML document.

Type Aliases

Array

The type contained in the Yaml::Array variant. This corresponds to YAML sequences.

Hash

The type contained in the Yaml::Hash variant. This corresponds to YAML mappings.

YAMLDecodingTrapFn

The signature of the function to call when using YAMLDecodingTrap::Call.